BIG WARNING:
If you want to build you will have to figure out the dependencies out yourself. I will get to this eventually.

*Some structural refactoring:

Finally got rid of the convoluted way in which I was registering delegates. Now it is very clean.

*Implemented initial Console system.

*Implemented initial Command system.

Using delegates for Console Command. Which is cool.

*Initial world loading.

Can load Collada scene files.