*** Removed Media folder. This is in preparation for reorganizing Media stuff due to the slight change to Media setup for Ogre 1.7.
****
Fixed bug (we need to setup bugzilla) which was causing crash on Ogre::Root destruction during auto_ptr destruction. Cause was in the destruction Ogreconsole code which I downloaded off the internet (was not deleting overlays properly.) We need to to move away from using Overlays (which OgreConsole uses) since I recall that Overlay usage is discouraged. 

Needs to fix the buffer overflow problem with OgreConsole still (switch to string tokening instead of looping through chars. Very yucky as of now.)