Fixed issue: http://code.google.com/p/projectzombie/issues/detail?id=5
Still try not to overwhelm it. By overwhelm I mean you move way ahead of the page loading, (you will start to see no handler processed request), by this time you are so far ahead there is no page around you and it takes awhile for the page to catch up. In normal usage of the app, this should be the exceptional case (you have to limit the max velocity of the user). 

Note, it still crashes sometimes, but now it crashes gracefully instead of access violation crash to desktop(i.e. the normal exception handling path is executed). And before, it would crash every time when "no handler processed request" start showing up in the logs, now, it won't crash on that. But instead, now it sometimes crashes with an Ogre movable object exception. So this points to the likelihood that the problem has been mostly fixed. Needs more testing.