-Fixed issue #1 with OgreConsole being crashy on exit. The problem arises from that we coupled it to CommandController which is supposed to be a singleton, and we tried to destroy the console after ogre shutdown. Another reason is because OgreConsole is a hack copy-and-pasta. We are planning to refactor Console into LibRocket anyway using all current system methodologies.
-System reboot on state change.
We implemented the feature where the entire engine can be torn down and restarted. This is good because it faciliate upgrading shaders at any time without shutting down the system. It's not complete yet because right now it will reload all the resources. In the future, we can divide say the main screen and the game. So there is less resource to load for main screen. Since configuration is tied to State, this can be easily done.