-More work on editor:
--Started work on transformation for tools. Implemented CubeScaleXForm. THe idea for this is that given any anchor position, it will scale a given node using scale operations only. Think of this as the typical rectangle
selection tool in a drawing program. This is done using scaling only, assume that the object's origin is at lower left side. Actually, this method will move the scene node that one is scaling relative to the anchor position.
Also note: although the function has been implemented, the wiring is not done yet, so it's not tested. See next.

--Started work on wiring scale mode. And thus we changed a few things on how the ToolsetViews* work.

--Changed ToolInfoView to update existing RML values for existing tools AND appending to for tool view changes.  

-Some other changes I don't recall 
--Maybe changed how values are generated where we're generating borders first. And only the bottom and right borders needs to be hit per volume chunk.